+++ site/homepage/content/release-notes.md
@@ -2,6 +2,14 @@
 title = "Release Notes"
 +++
 
+## 1.4.6 (not yet released)
+
+### Fixes & Enhancements
+
+* The protocol adapters erroneously indicated a client related error to devices if the downstream AMQP container
+  rejected a message with an `amqp:resource-limit-exceeded` error condition. This has been fixed so that the adapters
+  now correctly indicate a server related problem instead.
+
 ## 1.4.5
 
 ### Fixes & Enhancements